+++ packages/subgraph/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.1.2](https://github.com/cowprotocol/cow-sdk/compare/sdk-subgraph-v1.1.1...sdk-subgraph-v1.1.2) (2026-06-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/sdk-common bumped to 0.11.2
+    * @cowprotocol/sdk-config bumped to 2.3.0
+
 ## [1.1.1](https://github.com/cowprotocol/cow-sdk/compare/sdk-subgraph-v1.1.0...sdk-subgraph-v1.1.1) (2026-06-02)